requirements.txt
ollama
gradio
pdfplumber
langchain
langchain-core
langchain-ollama
langchain_community
langchain_text_splitters
langchain_openai
unstructured
unstructured[all-docs]
onnx==1.17.0
protobuf==3.20.3
chromadb==0.4.22
Pillow
numpy
urllib3<2.0
psycopg2
pgvector
python-dotenv
ocrmypdf
beautifulsoup4
playwright